pyproject.toml
[tool.poetry]
name = "minimal_telegram_bot"
version = "0.1.0"
description = ""
authors = ["Jonas Kemper <jonas.kemper.biz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^12.4.2"
requests = "^2.23.0"
pandas = "0.23.4"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
flake8 = "^3.7.9"
black = "^19.10b0"
isort = "^4.3.21"
pre-commit = "^2.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
